Regular Expression Fix and Added Globbing
I use matching groups out of habit, but it's not necessary.

Also looks like they're naming the DMG the same as the new file, and the Unarchiver processor has no Output Variables to reference. So I added a glob to the filename. Not sure if this will work but I'm trying it now.

Also attempting to add a new Input Variable. Could be useless but I'm all about that programmatic life.